requirements.txt
streamlit 
google.generativeai
json
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv